% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReduction.R
\name{GetReduction}
\alias{GetReduction}
\title{GetReduction}
\usage{
GetReduction(object, reduction, dims)
}
\arguments{
\item{object}{A Seurat object from which to extract the reduction}

\item{reduction}{The name of the reduction to extract}

\item{dims}{The dimension(s) of the reduction to extract. Defaults to the first two}
}
\value{
A matrix w/ reduction coordinates of the specified dims, per cell.
}
\description{
Extract the reduction from a Seurat object
}
\examples{
GetReduction(seuratobj, "umap", 1)
GetReduction(seuratobj, "pca", 1:2)
}
