% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomSubset.R
\name{RandomSubset}
\alias{RandomSubset}
\title{FindSubCluster}
\usage{
RandomSubset(object, by.factor, group.by = NULL)
}
\arguments{
\item{object}{A Seurat object to find subclusters in}

\item{cluster}{The cluster to be sub-clustered}

\item{graph.name}{The name of the graph to use for the clustering algorithm. Defaults to RNA_snn for convenience}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities.}

\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm). Leiden requires the leidenalg python.}
}
\value{
A Seurat object with subclusters stored in Idents(object) instead of a meta.data columns
}
\description{
Convenience wrapper for Seurat::FindSubCluster
}
\examples{
FindSubCluster(seuratobj, 'CD8+')
FindMarkers(FindSubCluster(seuratobj, "CD4+"), ident.1 = "CD4+_0", ident.2 = "CD4+_1")
}
